refactor: external services in different files
