+++ ex_app/lib/all_tools/weather.py
@@ -1,0 +1,37 @@
+# SPDX-FileCopyrightText: 2025 Nextcloud GmbH and Nextcloud contributors
+# SPDX-License-Identifier: AGPL-3.0-or-later
+import typing
+import urllib.parse
+
+import httpx
+from langchain_core.tools import tool
+from nc_py_api import Nextcloud
+
+from ex_app.lib.all_tools.lib.decorator import safe_tool
+
+
+def get_tools(nc: Nextcloud):
+	@tool
+	@safe_tool
+	def get_current_weather_for_coordinates(lat: str, lon: str) -> dict[str, typing.Any]:
+		"""
+		Retrieve the current weather for a given latitude and longitude
+		:param lat: Latitude
+		:param lon: Longitude
+		:return:
+		"""
+		res = httpx.get('https://api.met.no/weatherapi/locationforecast/2.0/compact', params={
+			'lat': lat,
+			'lon': lon,
+		},
+						headers={
+							'User-Agent': 'NextcloudWeatherStatus/ContextAgent nextcloud.com'
+						})
+		json = res.json()
+		if not 'properties' in json or not 'timeseries' in json['properties'] or not json['properties']['timeseries']:
+			raise Exception('Could not retrieve weather for coordinates')
+		return json['properties']['timeseries'][0]['data']['instant']['details']	
+
+	return [
+		get_current_weather_for_coordinates,
+	]
